Deployment config fixup.
This addresses the case where the directory name of a module doesn't
match the module's service name. We use the directory name in the
lists of modules at the top of the deploy configuration and at compile time. The
service name is used at runtime.

If a configuration block defines a key _service_name the DeploymentConfig code will
replicate the block containing that key into a block named by the value
of the _service_name. For example, if we have

[genome_annotation]
_service_name = GenomeAnnotation
name1 = value1
name2 = value2

this code will treat the configuration file exactly if it contained

[genome_annotation]
_service_name = GenomeAnnotation
name1 = value1
name2 = value2

[GenomeAnnotation]
_service_name = GenomeAnnotation
name1 = value1
name2 = value2